removed unused course admin controller from mac client. Fixed lots of compiler warnings on ios.

+++ source/ios/ImageHolderView.m
@@ -125,7 +125,7 @@
 	self.imageView.frame = imframe;
 }
 
-- (void)scrollViewDidEndZooming:(UIScrollView *)scroll withView:(UIView *)view atScale:(float)scale
+- (void)scrollViewDidEndZooming:(UIScrollView *)scroll withView:(UIView *)view atScale:(CGFloat)scale
 {
 	CGRect ourBnds = self.scrollView.bounds;
 	CGRect suggestedBnds = [self centeredFrameForScrollView:scroll andView:self.imageView];
